backend/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bce-python-sdk==0.8.83
bcrypt==4.0.1
beautifulsoup4==4.12.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.10.7
fire==0.5.0
Flask==2.3.2
flask-babel==3.0.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
flatbuffers==1.12
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.3
google-auth-oauthlib==0.4.6
gradio==3.23.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
imageio==2.27.0
imgaug==0.4.0
imutils==0.5.4
ipython-genutils==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
libclang==14.0.1
linkify-it-py==2.0.0
lmdb==1.4.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==0.4.3
networkx==3.0
numpy==1.24.2
oauthlib==3.2.0
onnx==1.12.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.8
packaging==23.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pdf2docx==0.5.6
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.1
premailer==3.10.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
PyMuPDF==1.20.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==306
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
rapidfuzz==2.13.7
rarfile==4.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.7
sqlparse==0.4.3
starlette==0.26.1
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
testpath==0.6.0
thop==0.1.1.post2209072238
tifffile==2023.3.21
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
tritonclient==2.31.0
typing_extensions==4.5.0
tzdata==2023.2
uc-micro-py==1.0.1
ultralytics==8.0.136
urllib3==1.26.15
uvicorn==0.21.1
visualdl==2.5.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.3.6
wincertstore==0.2
wrapt==1.14.1
x2paddle==1.4.0
yarg==0.1.9
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
